Let LDPC encoder support different code rates
Former-commit-id: 1b3dff747d5c480cb26c8ddaff327545f887b9b7

+++ src/common/Symbols.hpp
@@ -192,11 +192,11 @@ struct LDPCconfig {
     bool earlyTermination;
     int16_t decoderIter;
     uint16_t Zc;
-    int nRows;
+    size_t nRows;
     uint32_t cbEncLen;
     uint32_t cbLen;
     uint32_t cbCodewLen;
-    int nblocksInSymbol;
+    size_t nblocksInSymbol;
 };
 
 typedef struct LDPCconfig LDPCconfig;
